+++ CHANGELOG.md
@@ -22,7 +22,10 @@
 - Added error handling to search API
 - Refactored analytics report code
 - Implemented country-level analytics report
-- Implemented FRBRization v2 in OCLC Classify Process
+- Implemented total-usage analytics report
+- Added error handling to fulfill API
+- Added uuid API validation
+- Implemented OCLC Classify Process v2
 
 ## Fixed
 - Changed HATHI_DATAFILES outdated link in development, example, and local yaml files